pyproject.toml
[project]
name = "llm-playground"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.24.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
